requirements.txt
fastapi
uvicorn
chromadb
google-generativeai
python-dotenv